Removed ListView Redraw glitch
While moving through a ListView with the arrow keys redraw happened
correctly - but was not shown immediately. That way key strokes
sometimes seemed to happen delayed.

This was due to a Curses refresh happening too late. By adding the
Refresh() call right at the end of ListView.Redraw() the problem
vanished. Navigating the ListView items is now as expected.