Support for bindings to different ncurses version
Since I just removed all the glue code and native detection
dependencies as well as having a single binary for the curses support,
I had to write this support.

.NET Core sadly does not support Mono's <dllmap> yet, so it is not
possible to do this in any decent form.  This hell is imposed not only
on my curses binding, but on any library that needs to deal with Linux
packagin peculiarities - and we are inflicing this pain on every .NET
developer doing Unix variations.

Anyways, to make a long story short, now I have this ugly turd of a
binding to curses for the sole purpose of calling into one kind of
curses on MacOS, and another on Linux.  This could have been a 3 line
XML configuration file if we had supported dllmap, but instead this
has devolved into what could be described as poster child of bikeshedding.

This is one of that many issues, lots more where this came from:

https://github.com/dotnet/coreclr/issues/930